requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.10
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djongo==1.3.6
dnspython==2.4.2
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
numpy==1.24.4
packaging==23.2
pandas==2.0.3
PyJWT==1.7.1
pymongo==4.5.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.2.4
timedelta==2020.12.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
utils==1.0.1
